Simplify. Move related function from separate file.
