+++ src/ccstruct/ccstruct.h
@@ -22,11 +22,9 @@
 #include "ccutil.h" // for CCUtil
 
 namespace tesseract {
-class TESS_API CCStruct : public CCUtil {
-public:
-  CCStruct() = default;
-  ~CCStruct() override;
 
+class CCStruct : public CCUtil {
+public:
   // Globally accessible constants.
   // APPROXIMATIONS of the fractions of the character cell taken by
   // the descenders, ascenders, and x-height.
@@ -36,6 +34,7 @@ public:
   // Derived value giving the x-height as a fraction of cap-height.
   static const double kXHeightCapRatio; // = XHeight/(XHeight + Ascender).
 };
+
 } // namespace tesseract
 
 #endif // TESSERACT_CCSTRUCT_CCSTRUCT_H_
